update computers api add, rate limit on create

+++ packages/api/src/computers/mutations.ts
@@ -6,9 +6,19 @@ import {
   updateComputerParams,
 } from ".";
 import { IdType, idSchema, throwTRPCError } from "../common";
+import { createComputerLimit } from "../rate-limit";
+import { TRPCError } from "@trpc/server";
 
 export const createComputer = async (computer: InsertComputer) => {
   insertComputerParams.parse(computer);
+  const RL = await createComputerLimit.limit(computer.userId);
+  if (!RL.success) {
+    throw new TRPCError({
+      code: "TOO_MANY_REQUESTS",
+      message: "Rate Limit Exceeded, Try after few minutes.",
+    });
+  }
+  console.log("Remaining Limit: ", RL.remaining);
   try {
     const c = await db.computer.create({ data: computer });
     return { computer: c };
@@ -40,9 +50,10 @@ export const deleteComputer = async (id: IdType) => {
   }
 };
 
-export const deleteAllComputers = async () => {
+export const deleteUsersAllComputers = async (userId: IdType) => {
+  idSchema.parse(userId);
   try {
-    const c = await db.computer.deleteMany();
+    const c = await db.computer.deleteMany({ where: { userId } });
     return { computersDeleted: c.count };
   } catch (err) {
     throw throwTRPCError(err);
